Add alternative font sizes and mouse wheel scroll
Alternative font sizes are used in the CTRL+right click menu as well
Shift++/- (numpad) for switching fonts.

The alternateScroll function sends arrow up and down keys when a program
is displayed in the alternate buffer (which for example less uses) and
the mouse wheel is turned. This the mouse wheel can be used to scroll
through man pages (my main application)